DESeq2_test: differential expression analysis of RNASeq data using DESeq2

DESeq2_test: differential expression analysis of RNASeq data using DESeq2

Usage

DESeq2_test(
  data,
  column,
  groups,
  reference = "WT",
  covariates = c(),
  speedup = FALSE,
  parallel = FALSE
)

Arguments

data

data matrix of counts data

column

column name of the groups to be used for the definitions

groups

design matrix containing the definitions to be compared.

reference

Reference group to use in the DE analysis

covariates

List of column names in the design matrix to be used as covariates

speedup

Boolean parameter to indicate if the faster version of DESeq2 should be used with fitType = "glmGamPoi" (note: it may not work depending on the system)

parallel

Boolean parameter to indicate if DESeq2 should be run in parallel mode

Value

DESeq2 result object with the pvalues, adjusted pvalues and foldchanges of the DE analysis

Examples

# Internal function